调天线

+++ main/RFID/sky1311_drv.h
@@ -31,7 +31,7 @@ void sky1311ReadFifo(uint8_t *data, uint8_t count);
 
 void SPI_Config(void);
 void DelayMS(uint32_t x);
-void delay(volatile uint8_t i);
+void delay(volatile uint16_t i);
 void uart_puts(char *str);
 void uart_newrow();
 void uart_printBuffer(uint8_t* buff, uint16_t size);
